Revert POI to v.3.12

+++ DEV/org.openl.rules/src/org/openl/rules/lang/xls/XlsWorkbookSourceCodeModule.java
@@ -65,7 +65,7 @@ public class XlsWorkbookSourceCodeModule extends SourceCodeModuleDelegator {
         }
         Workbook workbook = getWorkbook();
         int numStyles = workbook.getNumCellStyles();
-        for (int i = 0; i < numStyles; i++) {
+        for (short i = 0; i < numStyles; i++) {
             CellStyle cellStyle = workbook.getCellStyleAt(i);
 
             wbColors.add(cellStyle.getFillForegroundColor());
